Long standing clean-up -- replaced "long long" and off_t in stemkit api calls with "uint64_t". Disambiguated reading and writing shorts, ints and longs by changing parameters and return values to uint16_t, uint32_t and uint64_t respectfully.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@1599 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/io/AbstractInputStream.h
@@ -30,7 +30,7 @@ namespace sk {
         std::vector<char> read(int number);
         std::vector<char>& read(std::vector<char>& buffer, int number);
         int skip(int number);
-        long long available() const;
+        uint64_t available() const;
         bool markSupported() const;
         void mark(int readlimit);
         void reset();
